Gikk tilbake til krypteringhandler i util, enighet om at å instansiere på klassenivå bør være ok
